Add ability to update publish url and install url
- Added the ability to pass in either the publish url or install
url and it will update the csproj file. If no parameter is provided
they will remain the same

- Switched around some right side null comparisons to left ones
due to warning encountered from VS Code and after doing a little research.
Found this article https://rencore.com/blog/powershell-null-comparison/

- Added tests for the publish url and install parameters

- Added test for using both the Version parameter and the BuildId
parameter. Passing in 1.0.0.0 as the Version and along with a BuildId
the version will keep the hardcoded major/minor(1.0) and update the
build/revision(0.0) with the autogenerated id